Add universal product system supporting finished products, raw materials, and accessories
- Add ProductType enum (FINISHED_PRODUCT, RAW_MATERIAL, ACCESSORY)
- Add UnitType enum (PIECE, METER, SQUARE_METER, KILOGRAM, ROLL)
- Update Product entity with new fields for raw materials and accessories
- Change quantity and minStockLevel from Integer to BigDecimal for decimal quantities
- Add V23 database migration for new columns and type changes
- Update ProductRequest/Response DTOs with new fields
- Add productType filter to ProductController and ProductRepository
- Update all services (Sale, Purchase, Stock, Report, Dashboard) for BigDecimal compatibility
- Add dynamic form in ProductsPage showing relevant fields based on product type
- Add PRODUCT_TYPES and UNIT_TYPES constants in frontend

+++ jalyuzi-epr-api/src/main/java/uz/jalyuziepr/api/service/DashboardService.java
@@ -48,7 +48,7 @@ public class DashboardService {
         BigDecimal todayRevenue = saleRepository.getTodayRevenue(startOfDay, endOfDay);
         BigDecimal totalRevenue = saleRepository.getTotalRevenue();
         long totalProducts = productRepository.countActiveProducts();
-        Long totalStock = productRepository.getTotalStock();
+        BigDecimal totalStock = productRepository.getTotalStock();
         long lowStockCount = productRepository.findLowStockProducts().size();
         long totalCustomers = customerRepository.countActiveCustomers();
         BigDecimal totalDebt = debtRepository.getTotalActiveDebt();
@@ -58,7 +58,7 @@ public class DashboardService {
                 .todayRevenue(todayRevenue != null ? todayRevenue : BigDecimal.ZERO)
                 .totalRevenue(totalRevenue != null ? totalRevenue : BigDecimal.ZERO)
                 .totalProducts(totalProducts)
-                .totalStock(totalStock != null ? totalStock : 0L)
+                .totalStock(totalStock != null ? totalStock.longValue() : 0L)
                 .lowStockCount(lowStockCount)
                 .totalCustomers(totalCustomers)
                 .totalDebt(totalDebt != null ? totalDebt.abs() : BigDecimal.ZERO)
